+++ pkg/state/release.go
@@ -58,7 +58,7 @@ func (r ReleaseSpec) ExecuteTemplateExpressions(renderer *tmpl.FileRenderer) (*R
 		ts := *result.WaitTemplate
 		resultTmpl, err := renderer.RenderTemplateContentToString([]byte(ts))
 		if err != nil {
-			return nil, fmt.Errorf("failed executing template expressions in release \"%s\".version = \"%s\": %v", r.Name, ts, err)
+			return nil, fmt.Errorf("failed executing template expressions in release \"%s\".waitTemplate = \"%s\": %v", r.Name, ts, err)
 		}
 		result.WaitTemplate = &resultTmpl
 	}
@@ -67,7 +67,7 @@ func (r ReleaseSpec) ExecuteTemplateExpressions(renderer *tmpl.FileRenderer) (*R
 		ts := *result.InstalledTemplate
 		resultTmpl, err := renderer.RenderTemplateContentToString([]byte(ts))
 		if err != nil {
-			return nil, fmt.Errorf("failed executing template expressions in release \"%s\".version = \"%s\": %v", r.Name, ts, err)
+			return nil, fmt.Errorf("failed executing template expressions in release \"%s\".installedTemplate = \"%s\": %v", r.Name, ts, err)
 		}
 		result.InstalledTemplate = &resultTmpl
 	}
@@ -76,7 +76,7 @@ func (r ReleaseSpec) ExecuteTemplateExpressions(renderer *tmpl.FileRenderer) (*R
 		ts := *result.ConditionTemplate
 		resultTmpl, err := renderer.RenderTemplateContentToString([]byte(ts))
 		if err != nil {
-			return nil, fmt.Errorf("failed executing template expressions in release \"%s\".version = \"%s\": %v", r.Name, ts, err)
+			return nil, fmt.Errorf("failed executing template expressions in release \"%s\".conditionTemplate = \"%s\": %v", r.Name, ts, err)
 		}
 		result.ConditionTemplate = &resultTmpl
 	}
@@ -85,7 +85,7 @@ func (r ReleaseSpec) ExecuteTemplateExpressions(renderer *tmpl.FileRenderer) (*R
 		ts := *result.VerifyTemplate
 		resultTmpl, err := renderer.RenderTemplateContentToString([]byte(ts))
 		if err != nil {
-			return nil, fmt.Errorf("failed executing template expressions in release \"%s\".version = \"%s\": %v", r.Name, ts, err)
+			return nil, fmt.Errorf("failed executing template expressions in release \"%s\".verifyTemplate = \"%s\": %v", r.Name, ts, err)
 		}
 		result.VerifyTemplate = &resultTmpl
 	}