+++ pkg/state/state_exec_tmpl.go
@@ -48,6 +48,10 @@ func getBoolRefFromStringTemplate(templateRef string) (*bool, error) {
 	return &result, nil
 }
 
+func setConditionFromBool(r *ReleaseSpec, condition bool) {
+	r.Condition = fmt.Sprintf("%t", condition)
+}
+
 func updateBoolTemplatedValues(r *ReleaseSpec) error {
 	if r.InstalledTemplate != nil {
 		if installed, err := getBoolRefFromStringTemplate(*r.InstalledTemplate); err != nil {
@@ -63,7 +67,7 @@ func updateBoolTemplatedValues(r *ReleaseSpec) error {
 			return fmt.Errorf("conditionTemplate: %v", err)
 		} else {
 			r.ConditionTemplate = nil
-			r.Condition = fmt.Sprintf("%t", *condition)
+			setConditionFromBool(r, *condition)
 		}
 	}